Project 1 Updates
Added methods to the character:
character.expGain(expAmount) - gives the character an amount of experience and displays the results in the combat screen
character.goldGain(goldAmount) - gives the character an amount of gold and displays the results in the combat screen
character.checkExp() - levels up the character if the desired amount of experience meets or surpasses the amount required to level, then displays the level (if one was obtained) message in the combat screen. Also increases the exp required for the next level by a multiplier of the character's level

Additional character stats:
character now includes race, gender and class
a boolean was added to show if the character is in combat

Fixed combat system:
enemies can no longer give multiple exp and gold rewards from a single death
Character can be killed. Once checkpoint system is implemented, it will be moved to that location

Stats display updates:
Connected stats display to gold, intelligence, race and gender of character. Also adjusted damage to use - between min and max instead of /. They now update accordingly